LayoutTests/platform/gtk/accessibility/custom-elements/describedby-expected.txt

This tests that aria fallback roles work correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.accessibleElementById("custom-1").role is "AXRole: AXCheckBox"
PASS accessibilityController.accessibleElementById("custom-1").helpText is "AXHelp: some description other description"
FAIL accessibilityController.accessibleElementById("custom-1").detailsElements().length should be 2. Threw exception TypeError: undefined is not an object (evaluating 'accessibilityController.accessibleElementById("custom-1").detailsElements().length')
FAIL accessibilityController.accessibleElementById("custom-1").detailsElements()[0].domIdentifier should be details1. Threw exception TypeError: undefined is not an object (evaluating 'accessibilityController.accessibleElementById("custom-1").detailsElements()[0]')
FAIL accessibilityController.accessibleElementById("custom-1").detailsElements()[1].domIdentifier should be details2. Threw exception TypeError: undefined is not an object (evaluating 'accessibilityController.accessibleElementById("custom-1").detailsElements()[1]')
FAIL accessibilityController.accessibleElementById("custom-1").errorMessageElements().length should be 2. Threw exception TypeError: undefined is not an object (evaluating 'accessibilityController.accessibleElementById("custom-1").errorMessageElements().length')
FAIL accessibilityController.accessibleElementById("custom-1").errorMessageElements()[0].domIdentifier should be error1. Threw exception TypeError: undefined is not an object (evaluating 'accessibilityController.accessibleElementById("custom-1").errorMessageElements()[0]')
FAIL accessibilityController.accessibleElementById("custom-1").errorMessageElements()[1].domIdentifier should be error2. Threw exception TypeError: undefined is not an object (evaluating 'accessibilityController.accessibleElementById("custom-1").errorMessageElements()[1]')
PASS accessibilityController.accessibleElementById("custom-2").role is "AXRole: AXCheckBox"
PASS accessibilityController.accessibleElementById("custom-2").helpText is "AXHelp: some description"
FAIL accessibilityController.accessibleElementById("custom-2").detailsElements().length should be 1. Threw exception TypeError: undefined is not an object (evaluating 'accessibilityController.accessibleElementById("custom-2").detailsElements().length')
FAIL accessibilityController.accessibleElementById("custom-2").detailsElements()[0].domIdentifier should be details2. Threw exception TypeError: undefined is not an object (evaluating 'accessibilityController.accessibleElementById("custom-2").detailsElements()[0]')
FAIL accessibilityController.accessibleElementById("custom-2").errorMessageElements().length should be 1. Threw exception TypeError: undefined is not an object (evaluating 'accessibilityController.accessibleElementById("custom-2").errorMessageElements().length')
FAIL accessibilityController.accessibleElementById("custom-2").errorMessageElements()[0].domIdentifier should be error2. Threw exception TypeError: undefined is not an object (evaluating 'accessibilityController.accessibleElementById("custom-2").errorMessageElements()[0]')
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
some description
other description
some details
other details
some error
other error
